Shade background of expression plots by CDC phase
Notebook shades background of expression plots (invoked by plot_genes()
and plot_replicates()) by CDC phase: G1 is red, S is blue, G2/M is
green.

Hard-coded the time points in a DataFrame (which is in turn derived from
a dictionary) because it isn't easy to grab the relevant time points
from sample_information for the following reasons:
- Time in minutes for each time point is not consistent.
- sample-information.csv doesn't explicitly list the time at which the
  cells transition from one phase to another (it would be impossible
  anyway, given that it is a population of cells).
So I took the latest time in minutes for the preceding phase, the
earliest time in minutes for the succeeding phase, and found the
midpoint.

Then, defined a dictionary to map CDC phases to a colour so that I can
conveniently put shading in a loop.