Copy of flavoprotein abundance with ribosome protein reference
Added the most abundant ribosome protein (RPS7A) so that bar plot will
include this one as a reference.  Additional column for colours so that
this protein is coloured differently.

Tech debt: notebook code as it is assumes that the proteins in the
spreadsheet to be imported are already sorted by abundance and thus
does not have lines to sort by abundance.